+++ Contest Prep Monolith.js
@@ -1,0 +1,3906 @@
+#include<iostream>
+
+
+https://codeforces.com/problemset/problem/227/B                // didn't understand the question
+
+
+
+https://codeforces.com/problemset/problem/1360/C             // question not crystal clear, how did sort like that.
+
+
+
+https://codeforces.com/problemset/problem/1490/C             //changing ll i=0 to i=1 makes all the difference.mpp
+
+
+
+https://codeforces.com/problemset/problem/1334/B             // sort(_);from the end sum+=a[i]; sum/nums>=x,count++,nums++; complex math.
+
+
+
+https://codeforces.com/problemset/problem/1345/B             //Binary Search,math,awesome question.
+
+
+
+https://codeforces.com/problemset/problem/1324/B             // 3d array.
+
+
+
+https://codeforces.com/contest/1657/problem/A                   //isPerfectSquare(z), floor(z), ceil(z).
+
+
+
+https://codeforces.com/contest/1657/problem/B                   // didn't understand the question.
+
+
+
+https://codeforces.com/problemset/problem/1472/C            // recursion.
+
+
+
+https://codeforces.com/problemset/problem/1382/B             // continue;
+
+
+
+https://codeforces.com/contest/349/problem/A                      // in a[i]==100, what if if(twofive>=1&&fifty>=1) and elseif(twofive>=3) both get executed.
+
+
+
+https://codeforces.com/problemset/problem/1342/B            //change "" to '' what happens. ll first=s[0]-'0'.UNderstood the code but not the question.
+
+
+
+https://codeforces.com/contest/1657/problem/B                   // yet another question not comprehensible.
+
+
+
+https://codeforces.com/problemset/problem/1333/B             // check that last testcase && whatIf 2 2 0
+                                                                                                                                                                2 2 2
+
+
+
+https://codeforces.com/problemset/problem/1370/B             //compilation error, where is (Remove any two elements from a and append their sum to b), If there are multiple answers you can find any.vector<pair<ll,ll>>
+
+
+
+https://codeforces.com/problemset/problem/478/A               //what about "a coin is passed from one player to some other player" question and answer entirely different from each other.
+
+
+
+https://codeforces.com/problemset/problem/1213/B              //if you do i++ after(a[i]>a[j]) i will become i+=2. analyse all 3 submissions.
+
+
+
+
+https://codeforces.com/problemset/problem/1656/B              //BINARY SEARCH, without sort-wrong, with sort-accepted, check both //'s
+
+
+
+https://codeforces.com/problemset/problem/1481/B              //break doesn't break form the ifElse loop but the WhileFor loop
+
+
+
+https://codeforces.com/contest/1455/problem/A                    //xz(s)...xD Look how puzzling it looked and how silly it was.
+
+
+
+https://codeforces.com/problemset/problem/1631/B             // whats the use of if(i<0)break when thw while loop already has the same condition, i-=cnt before cnt*=2;
+
+
+
+https://codeforces.com/problemset/problem/1626/B              //substr, continue, same code diff output.
+
+
+
+https://codeforces.com/problemset/problem/1511/C             //map<colour,index>
+
+
+
+https://codeforces.com/problemset/problem/1593/D1            // __gcd() to make all numbers equal to the lowest number.
+
+
+
+https://codeforces.com/problemset/problem/1553/C              //RECURSION easy if you know    PS:notSolved
+
+
+
+https://codeforces.com/problemset/problem/1324/C               // Frog Jumps, char[2000], strlen(), i-previndex, mistake (i==n), d cannot be changed as in question but testcase 3 d changed?? previndex has to be -1,else WA
+
+
+https://codeforces.com/problemset/problem/1293/A                // BINARY SEARCH, map and set used in subsequent submissions
+
+
+https://codeforces.com/problemset/problem/1221/C                // Hardcore Math
+
+
+
+https://codeforces.com/contest/1472/problem/B                       // taking indexing from zero gives wrong answer
+
+
+
+https://codeforces.com/contest/1472/problem/A                       //RECURSION 1st problem
+
+
+
+https://codeforces.com/problemset/problem/1042/A                //didn't understand
+
+
+
+https://codeforces.com/problemset/problem/919/B                 //int sum(), i<11000000 (every zer0 taken seriously)
+
+
+
+https://codeforces.com/contest/1006/problem/C                     //2 pointers
+
+
+
+https://codeforces.com/contest/755/problem/B                        //set used to erase duplicates.2nd submission, does set only store unique values?
+
+
+
+https://codeforces.com/contest/743/problem/B                       //pow()?? Refer all submissions.Sanatorium code passed the
+                                                                                                                                                               first testcase,lol.
+
+
+https://codeforces.com/contest/637/problem/B                       //MAP string, to find if it is repeated in the array.
+
+
+
+https://codeforces.com/contest/735/problem/B                       //Question??, setprecision(8),giving weird answers
+
+
+
+https://codeforces.com/contest/729/problem/B                         //Hradcore DP.
+
+
+
+https://codeforces.com/problemset/problem/389/B
+MATRIX, Rows->Horizontal->i, Columns->Vertical->j,[ i ][ j ]
+
+
+
+https://codeforces.com/problemset/problem/302/B                  //Hard For ME
+
+
+
+https://codeforces.com/problemset/problem/285/C                  //Permutations
+
+
+
+https://codeforces.com/problemset/problem/262/B                  //Some details in code not in the question, sublime not processing testcase 5, mn != INT_MIN  mn=100000
+
+
+
+https://codeforces.com/problemset/problem/545/B                  //Question not understood
+
+
+
+https://codeforces.com/problemset/problem/1506/B                //where does the loop break in submission 4. why j=1 in the accepted submission.
+
+
+
+https://codeforces.com/contest/1667/problem/A                       // Question not clear, #define INT_MAX 1e18, overflow in 3 submissions.
+
+
+
+https://codeforces.com/contest/1668/problem/A                       // review this once
+
+
+
+https://codeforces.com/contest/1659/problem/A                         //if you have B balls and you want partitions such that you get the least consecutive R's, partitions=b+1
+                                                                                                              check the code(string ss,    i<b+1-Bleft,    s.pop_back()), different answers accepted
+
+
+https://codeforces.com/contest/1665/problem/A
+Just like this question the output given may not be the necessary answer so read the question.
+
+
+
+
+https://codeforces.com/contest/1566/problem/C
+using vector<pair<char,char>>, for 10 MEX is 2 but for 01 also is the MEX 2??,checking if i+1<n and using i++ wisely
+                                                                                                                                                                          10                             01
+
+
+
+https://codeforces.com/contest/1566/problem/B
+ s.erase(unique(s.begin(),s.end()), s.end());for erasing consecutive duplicates,     MEX cannot me more than 2
+
+
+
+
+
+https://codeforces.com/problemset/problem/1593/C
+ look for the binary search approach
+
+
+
+
+
+https://codeforces.com/problemset/problem/1472/C
+ starting the loop from n-1 and smartly changing the vector values to know the jumps further [DP]
+
+
+
+
+
+https://codeforces.com/contest/1469/problem/B
+starting from index 1, adding the number before and changing the value in the question vector. Easy code difficult question
+                        then nested loop to find the greatest possible sum
+
+
+
+
+
+https://codeforces.com/problemset/problem/1419/D1
+sorting the q array, keeping the value of j and forming an array with smaller number between 2 greater numbers[see code].
+
+
+
+
+
+https://codeforces.com/contest/1613/problem/A
+using pop_back(), then comparing the size of the numbers to check which is greater
+
+
+
+
+
+https://codeforces.com/contest/1613/problem/B
+why loop till n/2, testcases not matching but still accepted
+
+
+
+
+https://codeforces.com/contest/1551/problem/A
+math,coins in the form c1+2c2 so if 2 remainder after dividing by 3, 1 will be added to 2c2(2*1), got to know if you don't declare the value to zero the output is a huge number
+                                                                                                                                                                     and if 1 remainder                                   1 will be added to c1(1*1)
+
+
+
+
+https://codeforces.com/contest/1593/problem/B
+j-i-1 to remove numbers between i and j. (n-1)-j to remove all the numbers on the right of j
+min(ans,j-i-1+(n-1)-j) to remove the least possible numbers in order to make it divisible by 25
+
+
+
+
+https://codeforces.com/contest/1606/problem/B                             The approach of extreme char to be same.
+
+
+
+
+
+https://codeforces.com/contest/1606/problem/B                             understood until number of computers updated was smaller than the number of cables.
+                                                                                                                formula (remaining + k-1)/k for time taken by remianing computers.
+
+
+
+https://codeforces.com/contest/1569/problem/A
+using k pointer to start from index zero and keeping it between i and j(j started from i+1)
+ accepted without testcases matching.
+check the submissions.
+
+
+
+https://codeforces.com/contest/1569/problem/B
+ HOTS
+
+
+
+
+https://codeforces.com/contest/1559/problem/A
+    AND &
+
+
+
+
+
+https://codeforces.com/problemset/problem/550/A
+upper_bound            ABABAB
+
+
+
+
+
+https://codeforces.com/contest/1722/problem/C
+in 2d vector i is the vertical index, v[i] is used to denote the vector in the ith line  ||||  using auto x:v[i] x pointing to the index iterating over the array in the ith line.
+
+
+
+
+
+https://codeforces.com/contest/1682/problem/B
+ BITWISE AND, initializing ans to n int ans=n? ans=misplaced[0]?     no output if size of vector is not declared.
+
+
+
+
+
+
+https://codeforces.com/contest/1682/problem/C
+Revise both submissions
+
+
+
+
+https://codeforces.com/contest/1700/problem/A
+shortest path in matrix, ap, gp FORMULA
+
+
+
+
+
+https://codeforces.com/contest/1695/problem/A
+i,n,x->row(x axis)   j,m,y->column(y axis)       x=i+1 and n+1-x     +1 is because the indexing is from zero
+
+
+
+
+
+https://codeforces.com/contest/1681/problem/B
+%n card shuffle    these kind of approaches is What I want to master. Got it by using pen and paper.
+
+
+
+
+
+https://codeforces.com/contest/1681/problem/C
+break immediately after the ans vector is not how it should be after any iteration.
+
+
+
+
+
+
+https://codeforces.com/contest/1670/problem/B?csrf_token=0853d84713c19fc2a2baa8d0d3afd5da&f0a28=1
+i<26  { v[i]=0; }       cin>>x  v[x - 'a'] = 1    if(v[s[i] - 'a'] != 0)
+                                                                                                                          TLE
+
+
+
+https://codeforces.com/contest/1674/problem/A
+testcases not matching but still accepted (b^a)*x=y so if y is divisible by x print and when a=1, b=y/x, therefore 1<<y/x<<\n
+if y is not divisible by x then its not possible
+
+
+
+https://codeforces.com/contest/1674/problem/B
+our indexing  ba  bb  bc      what testcases want  ba  bb  bc
+                                                                                                                        0     1    2                                              1           2
+
+
+
+
+https://codeforces.com/contest/1674/problem/C
+learned to output power      (long long) pow(base, exponent)
+
+
+
+
+
+
+https://codeforces.com/contest/1695/problem/C
+adding pairs in a matrix, DOUBT       [if(a[n-1][m-1].first <=0 && a[n-1][m-1].second >=0 && (n+m-1)%2==0)]
+Using make_pair
+
+
+
+
+
+https://codeforces.com/contest/1684/problem/C
+ SEE CODE to understand how to smartly use vector<vector<int>>v(n, vector<int> (m))
+                                                                                        auto b=v[x];
+
+
+
+
+
+https://codeforces.com/contest/1695/problem/D1
+Tree, DFS implementation, exuberant for Dsa
+
+
+
+
+https://codeforces.com/contest/1673/problem/A
+s[i]-'a' +1         like does a have value 0, b 1, c 2, d 3....z 25. BUT in the question z 26 so +1 for every character.
+
+
+
+
+
+https://codeforces.com/contest/1673/problem/B
+initializing the value of index to s.size() instead of n changes the output, s[i] != s[i%index] to check if the difference is more than one in the substring
+                                                                                                AWESOME SUBSTRING question    map<char,int>
+                                                                                                BUT it will fail for testcase abcacb
+
+
+
+
+
+https://codeforces.com/contest/1673/problem/C
+ COIN CHANGE problem, string s = to_string(i),  table[40005], define MOD 1000000007,  table[j] %= MOD;
+
+
+
+
+
+https://codeforces.com/contest/1672/problem/A
+by observing, we came to the conclusion that no matter how you cut the log there will be a-1 operations where a is the length of the log
+
+
+
+
+https://codeforces.com/contest/1672/problem/C
+by observing we came to the conclusion that to make the equality of a string of size n, 1 we need n-3 operations, max(1, n-3) is used for edge case 111
+
+
+
+
+
+
+https://codeforces.com/contest/1672/problem/D
+using multiset, look at the code for the i-- doubt, and how will index be equal to zero when the indexing is from zero
+instead of set                            using CONTINUE
+
+
+
+
+
+https://codeforces.com/contest/1671/problem/B
+To check if it can be consecutive if a[i] can be increased or decreased by 1,  (last element - first element) - n,  should be less than 1
+
+
+
+
+https://codeforces.com/contest/1654/problem/B
+map usage
+
+
+
+https://codeforces.com/contest/1654/problem/C
+at any point when the number of elements in the queue is greater than the remaining number of elements in the array, break:CONCEPT
+insert() in multiset
+push() in queue
+ceil() and floor() ceil didn't work :(
+so abc/2 + abc%2 alternative for ceil
+multiset is usually used where you have to find() or erse an element()
+
+
+
+
+
+https://codeforces.com/contest/1625/problem/A
+converting decimal to binary
+then converting binary to decimal
+adding binary numbers, Storing the value of the number of 1's by adding the ones place, tenths place hundreths place and thousandths place
+
+
+
+
+
+
+
+https://codeforces.com/contest/1625/problem/C
+To realise if its a dp problem, you see the constraints n<=500, k<=n-1 so a code of n^3 can also be written
+
+
+
+
+
+
+
+https://codeforces.com/contest/1627/problem/A
+Points to Remember:
+map<int,int> row, col;  /*     Making 2 maps if there is
+row[i]=1;                      a 'B' in that row then 1 or
+col[j]=1;                      if there is a 'B' in that column then 1; *\
+if(row[r]==1 || col[c] == 1)// to check if there is a 'B' in the rook(chess) direction of the cell
+mistake #1 int[n+1][m+1] instead of char[n+1][m+1]
+mistake #2 if(), if(), if() instead of else if(), else if(), else()
+mistake #3
+
+
+
+
+
+https://codeforces.com/contest/1608/problem/B
+This question was so hard that this was PRIYANSH31DEC's worst contest;
+we came up with 3 conditions by observing that the absolute difference of the number of maximas and minimas should be less than or equal to 1 i.e. |a-b| <= 1;
+Peak and Valley approach; for(a==b+1){ 1 more peaks than valleys }
+                          for(a==b-1){ 1 more valley than peaks }
+
+
+
+
+https://codeforces.com/contest/1621/problem/B
+map<pair<ll,ll>, ll> mpp;
+if(mpp.find(make_pair(l,r)) != mpp.end())
+{
+  mpp[make_pair(l,r)] = min(c, mpp[make_pair(l,r)]);      if the new input l and r already exist in the map as a pair we check which one has the lower cost
+}
+else
+{
+   mpp[make_pair(l,r)] = c;
+}
+
+if(mpp.find(make_pair(minl,maxr)) != mpp.end())
+{
+  ans = min(ans, mpp[make_pair(minl,maxr)])       if the MINl and MAXr for which we have calculated the ans=costl+costr already exists as a pair in the map we check which one has the lower cost;
+}
+
+
+
+
+
+https://codeforces.com/contest/1624/problem/A
+Understanding the question was the real deal here
+
+
+
+
+https://codeforces.com/contest/1624/problem/B
+Arithmetic Progression
+c-b=d so c=b+d----(1)
+b-a=d so a=b-d----(2)
+by(1) and (2)
+a+c=2b
+2b-c = a
+2b-a = c
+so to find if a can be multiplied to make it an AP we do(2b-c)%a will be equal to 0 if it is possible to make an AP
+also using else if() here
+
+MISTAKE #1 whenever multiplying use brackets to separate 2 like (a+c)*(2*b);
+
+
+
+
+
+
+https://codeforces.com/contest/1624/problem/C
+using set to store unique values
+keep dividing the number by 2 until it becomes <= n and a similar value is not present in the set
+if the number becomes zero the answer will be no
+
+
+
+
+
+https://codeforces.com/contest/1702/problem/A
+USING for(i=1; i<1e10; i*=10);
+
+
+
+
+https://codeforces.com/contest/1702/problem/B
+Using set to store unique char;
+MISTAKE intitalizing a=0, should be a=1 because the final 3 chars will not be included otherwise;
+
+
+
+https://codeforces.com/contest/1702/problem/C
+using 2 maps map<int,int> left, right;
+if(!left[v[i]])
+{
+  left[v[i]] = i+1;
+}
+if(left[l] && right[r] &&)            To check if l and r EVEN exist in the input array;
+
+
+
+
+
+
+https://codeforces.com/contest/1714/problem/C
+stoi()
+for(int mas = 0; mas<1<<9; mas++)
+
+
+
+
+
+
+
+https://codeforces.com/contest/1714/problem/A
+for(int i=0; i<N; i++)
+     {
+       cin>>H>>M;
+       int alarm = 60*H+M;
+       ans = min(ans, (alarm + 1440 - sleep) % 1440);
+     }
+cout<<ans/60<<" "<<ans%60<<endl;
+
+
+
+
+
+https://codeforces.com/contest/1714/problem/B
+MY SUBMISSION:
+using std::map<key, value> map;
+HIS SUBMISSION:
+vector<bool> seen(n+1, false);
+for(auto &x:v)
+{
+  cin>>x;
+}
+while(end>0 && !seen[v[end-1]]) ////////
+      {
+        end--;
+        seen[v[end]] = true;
+      }
+
+
+
+
+https://codeforces.com/contest/1702/problem/D
+*HARD
+where to break
+
+map<int,vector<int>> mpp;
+
+       int score = 0;
+
+       for(int i=0; i<w.size(); i++)
+       {
+         score += w[i]-'a' + 1;              adding the value for eg a=1, b=2, d=4.....z=26;
+         mpp[w[i]-'a' + 1].pb(i);
+       }
+
+        while(score>p)
+        {
+           for(int i=26; i>=0; i--)
+           {
+              for(auto it:mpp[i])            iterating over the vector on first index in the map
+              {
+                score-=i;
+                w[it] = '?';
+                if(score<=p)
+                {
+                  break;
+                }
+
+              }
+              if(score<=p)
+                {
+                  break;
+                }
+            }
+        }
+string ans = "";        declare empty string
+for(auto it:w)          iterating over the string
+{
+  if(it != '?')
+  ans +=it;             pushing that char in the array
+}
+
+
+
+
+
+
+https://codeforces.com/contest/1713/problem/B
+Check if it is decreasing;
+if(yes)
+then check if it is increasing further;
+if(yes)
+then print no
+
+
+
+
+
+
+
+https://codeforces.com/contest/1713/problem/C
+IMPORTANT NOTE:
+if the size of the new array is 5 and you have not taken it as an input using i=1 loop then by default the indexing is 0 1 2 3 4
+SO if you cout<<array[array.size()] a garbage value will gfet printed
+-> In this question even if n is 6 we are taking maxi=n-1=5 because our square vector also going to have 0 i.e {0 1 4 9}
+ALGORITHM
+1)KeepinG the ptr on the last element of the square vector.
+2) subtratcting i from the square[ptr] to get the element that will be pushed in the ans vector.
+3)If that element exceeds n or that element has alreday being pushed then we decrement the pointer which takes us to the second last element in the square vector
+  and now the next element is found by subtracting i from the new square[ptr];
+4)We insert element in the ans vector such that the sum of the element and the index in which it is inserted is a perfect square i.e. ans[element]=i; element - 12  12+13=25 which is a perfect square
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1716/problem/A
+The edge case where n==1
+The case where n%3==1 the optimal way is taking first 2 stops as 2 instead of 3 then at the end 3 will be left so do +1;
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1716/problem/B
+Changing the vector size to n+1, n+2
+DOUBT: Why does it give a wrong output when I follow 1 based indexing;
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1711/problem/B
+CONCEPTS:
+Adjacency List          g[x].pb(y),g[y].pb(x)   int x=it.first;int y=it.second
+Graphs                                                                                                                                                                     index - 13
+Changing the return 0 after case 1 to continue otherwise it gave only onw output that is zero
+
+CASE 1:
+if number of pairs is even
+number of cakes will also be EVEN
+so invite everyone
+unhappiness value will be zero
+
+CASE 2:
+If number of pairs is odd:
+Then we will remove 1 person, so even pairs will remain on condition that 1 person that will be removed forms odd number of pairs.
+SUBCASE:
+But what if every single person forms even pairs only then we will remove a pair such that both members of that pair form even pairs individually
+NOT UNDERSTANDING WHILE REVISING:
+So here we use the g vector to see if the number of pairs is even or odd
+CASE 2 in the comments represents the SUBCASE.
+DOUBT:
+I didn't understand can we overwrite in an array like g[x].pb(y) && g[y].pb(x) if we have{(1,3),(1,5),(1,4)}
+and how is if(g[i].size()%2==0) possible in a 2d vector
+
+
+
+
+
+
+
+https://codeforces.com/contest/1684/problem/A
+Observation based
+EDGECASE:
+when the string has only 2 digits answer will be the second number and not the lowest numbers
+
+
+
+
+
+
+https://codeforces.com/contest/1684/problem/B
+Solving the equations, refer how the BITS' guy (code explainer) came up with those equations and then said "it comes with practice";
+
+
+
+
+
+
+https://codeforces.com/contest/1711/problem/C
+ODD-ODD=even
+REFER the comments on code to Understand
+rows->n, columns->m  BUT columns that can be coloured = a[i]/n          rows that can be coloured = a[i]/m;
+changing all return 0  to  continue;
+Dealt with runtime error by declaring array a[N] where big N = 1e5+5
+Dealt with "wrong answer on test 8" by declaring long long columns_coloured and long long rows_coloured.
+
+
+
+
+
+https://codeforces.com/problemset/problem/339/A
+one line python code
+#REDO
+
+
+
+
+
+
+https://codeforces.com/contest/1735/problem/A
+Undertsanding the math behind it made the code Easy
+#REDO
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1729/problem/C
+using a map<char,vector<int>>   to store the index of the frequencies of an alphabet
+using the default indexing of the letters by char c = s[0]+i;
+for(auto x:mpp[c]) //used to iterate over the vector corresponding to the char c in the map;
+
+REFER the code comments once while revising.
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1729/problem/D
+In the solution we have done maximum 2 people go to the restaurant together but in the question more than 2 can go BUT by observing, 2 seems to be the most optimal solution
+REFER brocoders video to know how even 2 at a time will give us the same answer as more than 2 at a time.
+
+
+
+
+
+https://codeforces.com/contest/1739/problem/A
+Chess
+
+
+
+
+
+
+
+https://codeforces.com/contest/1739/problem/B
+Without the AMPERSAND & in (auto &it:v) the output was 0 0 0 0 0.
+REFER code comments
+#REDO for better understanding of questions like these
+
+
+
+
+
+
+https://codeforces.com/contest/1734/problem/A
+how to find the lowest difference between 3 elements in an array using SORTING.
+
+
+
+
+
+
+
+https://codeforces.com/contest/1734/problem/B
+UNDERSTANDING the question to know it wants all the rooms in a line to have the same amount of light, so light only required in the extreme rooms.
+POSITIONING of the statement cout<<endl;
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1730/problem/A
+My first submission failed because the code was not considering the n-1th element
+THERE WAS A CATCH IN THIS QUESTION:
+Since we are storing the frequency on the index of the vector the size of the vector should be 100 because the maximum number of planets that can be present in a galaxy is 100
+This submission using map
+
+
+
+
+
+
+
+https://codeforces.com/contest/1730/problem/B
+IF you don't specify the size of the vector, you don't get the output
+#cout<<setprecision(10)  without this it will give wrong answer.
+v[i]+t[i] gives us the total time taken by the person
+v[i]-t[i] gives us the position of the house of the person
+so we are actually doing [position of the leftmost house(mini) + position of the rightmost house(maxi)]/2
+We are assigning the leftmost and rightmost position on the basis of TIME.
+#tricky
+
+
+
+
+
+
+
+https://codeforces.com/contest/1730/problem/C
+The main reason this is a 1200 lvl problem is why we are iterating from n-1 is because we are sending the first occurence of the largest number from the back to the front in order to get the lexicographically smallest number out of the available numbers
+like 12215 iterating from back will give 11533 and from front will give 11336
+using (s[i]<=mini) instead of (s[i]<mini);
+#tricky
+
+
+
+
+
+https://codeforces.com/contest/1733/problem/A
+EASY:
+use map to store number with the corresponding i%k and keep changing as you find a greater number at that i%k
+use it.second to iterate over the numbers stored and print their sum.
+
+
+
+
+
+
+https://codeforces.com/contest/1733/problem/B
+These loops for giving the output:
+for(int i=0; i<n-1; i+=x)
+       {
+         for(int j=0; j<x; j++)
+         {
+These conditions used:
+if(!flag)
+player+=x+1
+flag = true;
+else
+player+=x;
+
+REFER Comments of the code
+
+
+
+
+
+
+
+https://codeforces.com/contest/1734/problem/C
+COME back to this #REDO
+
+
+
+
+https://codeforces.com/contest/1703/problem/B
+using a map<char,ll>
+for(char it:s) //for iterating over the string
+
+
+
+
+
+
+
+https://codeforces.com/contest/1703/problem/C
+#DOUBT why when v[i]=='D' we did cnt++ but 'D' means we are decreasing
+MAYBE because we use the statement v[i]=(v[i]+cnt+mod*10)%10;
+IN THE SECOND SUBMISSION:
+instead of using the MOD statement if(v[i]==-1)v[i]=9; if(v[i]==10)v[i]=0;
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1703/problem/D
+using v[i].substr(0, j);
+using map<String,bool>;
+using cout<<(flag?:0) stands for if(flag)cout<<1;else cout<<0;
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1703/problem/E
+vector<vector<int>>v(n, vector<int>(n)); for a MATRIX
+in other questions n->i->row->horizontal, but here n?i->column->vertical;
+Checking coordinates after flipping the matrix. In first box always the vertical and in the second always horizontal[vertical][horizontal] after rotation i or j become n-1-i or n-1-j when we look for uptodown and righttoleft //check code
+check COMMENT
+cout<<ans/4??
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1703/problem/F
+#DOUBT
+Didn't understand the logic for ans+=v[i]<i;
+also pref[i] = pref[i-1] + (v[i]<i); #PREFIX SUM
+Note that some indexes of the pref vector are 0 as we are checking if v[i]<i and only then adding it to the prefix. 0 because we declared vector<int> pref(n+1, 0);
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1741/problem/A
+INthe final statement we have compared the ascii values
+a new kind of while loop while(a[i]=='X')xa++, i++; and then using (a[i]==b[j]) to compare the last characters
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1741/problem/B
+OBSERVATION:
+if n is even then the index and the element will never be the same
+we are printing it from i=n because index and element will converge if we output from 1
+we found out a mid point where the index and element will converge in the case where n is odd.
+CAREFULLY use i<=mid, i>mid, i>0 etc.
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1741/problem/C
+using pointers
+look at the otehr binary search approach(brocoders)
+
+
+
+
+
+
+
+https://codeforces.com/contest/1690/problem/C
+refer comments.
+comparing the previous ending to the current starting to know which is greater
+
+
+
+
+
+
+https://codeforces.com/contest/1690/problem/D
+TLE if the size of Bsofar(n)
+int n, k;
+       cin>>n>>k;
+       string s;
+       cin>>s;
+
+       vector<int> Bsofar(n+1);
+
+       for(int i=1; i<=n; i++)
+       {
+           Bsofar[i] = Bsofar[i-1];
+           if(s[i-1]=='B')
+           {
+             Bsofar[i]++;
+           }
+       }
+
+      int len = k;
+      for(int i=k; i<=n; i++)
+      {
+        len = min(len, k-(Bsofar[i]-Bsofar[i-k]));
+      }
+
+      cout<<len<<endl;
+#PREFIXSUM
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1690/problem/E
+first add cost a[i]/k to the ans and then change it to the remaining a[i]%k;
+sort a.
+put 2 pointers and add the 2 if the answer is a[start]+a[end]>=k increment ans by 1.
+if it is less than k then move to a larger value i.e. start++;
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1742/problem/C
+it got accepted but didn't show all outputs n sublime.
+the logic is if red is coloured last we will get a complete row of RRRRRRRR.
+
+
+
+
+
+
+
+https://codeforces.com/contest/1742/problem/D
+in the ind vector storing the index of the number as the element and the element as the index.
+then using nested loops checking if those i and j are numbers even exist in the array and also checking if the 2 elements are co-prime using if(ind[i] && ind[j] && __gcd(i,j)==1)
+
+
+
+
+
+
+
+https://codeforces.com/contest/1204/problem/A
+#REDO
+https://codeforces.com/contest/339/problem/A
+In this question you make a new array and add all the numbers to it, sort it, and then use a pointer to replace the numbers in sorted order leaving the '+' signs untouched using s[i]=ouput[ptr++];
+
+
+
+
+
+
+
+https://codeforces.com/contest/1675/problem/A
+BRUTEFORCE
+deducting food,
+if dogs are left,
+deducting common food,
+if still left,
+no
+
+
+
+
+
+
+
+https://codeforces.com/contest/1675/problem/C
+HOTS:
+We observed that the suspects will be between the last occurence of 1 and the first occurence of 0.
+if
+nobody said 1
+then
+everybody before the first 0 including that person(here no +1 required because storing the index already includes this person)
+if
+nobody said zero
+Then
+everybody after the last 1 including that person and so that +1 in n-last_one+1 (+1 also because of the case "??11")
+
+
+
+
+
+
+https://codeforces.com/contest/1650/problem/B
+#REDO
+#MATH
+For better understanding of the math behind this.
+
+
+
+
+
+https://codeforces.com/contest/1650/problem/C
+the output is the index of the least coordinate with the index of the greatest co-ordinate.
+we shorlisted these co-ordinates on the basis of the the first 2*n coordinates of the sorted weight vector.
+#REFER code comments.
+
+
+
+
+
+https://codeforces.com/contest/1637/problem/A
+#DOUBT
+why does it give me a wrong output if I initialise b(n);
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1637/problem/C
+initialising v(n) gives RUNTIME ERROR
+Impossible case 1:
+all are 1's
+Impossible case 2:
+there are 3 numbers(n==3) and the middle number is odd.
+All other cases give an answer that is:
+ans+=(a[i]+1)/2.
+
+
+
+
+
+
+https://codeforces.com/contest/1520/problem/B
+BRUTAL MATH
+implementation like that
+(i=1;i<=9)
+{
+  start=i
+  while(start<=n)
+  start*=10 + i    ordinary numbers are 1,2,3,4,5,6,7,8,9,11,22,33,44,55,66,77,88,99
+}
+
+
+
+
+
+
+
+https://codeforces.com/contest/1520/problem/C
+OBSERVATION:
+edgecase(n==2)
+not possible
+OBJECTIVE:
+No adjacent numbers close to each other
+So, first fill the ans vector with all odd numbers till n^2
+then fill the vector with all even numbers till n^2
+OUTPUT:
+as we have to make a nXn matrix so whenever (i%n==0) cout<<endl;// check code
+we also use v[i-1] instead of v[i] because the declared vector has 0 based indexing by default.
+
+
+
+
+
+https://codeforces.com/contest/1646/problem/A
+Get back to the explanation by the IIT-Delhi guy.
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1646/problem/B
+The two-pointer approach was far easier than the other approach by the IIT guy
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1642/problem/A
+y2==y3,y2==y1,y3==y1 are the y coordinates of the line that is parallel to the x axis and is the unsafe border which means it is inaccessible from any vertex below this line
+and the length of this line is asked which is abs(x2-x3), abs(x2-x1), abs(x3-x1).
+
+
+
+
+
+
+
+https://codeforces.com/contest/1642/problem/C
+This question was so fucked up even the submission i copied and pasted didn't get accepted
+I finally had to copy the exact same thing as it is and submit in order to
+#REDO
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1649/problem/A
+we got j-i+2 equation by solving the equations:
+(j+1)-(i-1)
+we got j+1 because j has come over a 0 which means it has come one step behind more and in a similar way i has come one step more forward so it has to go one step back so i-1
+
+
+
+
+
+
+
+https://codeforces.com/contest/1649/problem/B
+if the maximum element is more than the sum of the rest then that dude will play with one ball until number of passes=sum with the rest of the people then for the leftover passes he will need as many balls as the leftover passes
+by our(brocoders) OBSERVATION if the sum is greater or equal to the largest element then only 1 ball will be required always as the highest dude or dudes will play
+alternatively with all the remaining players
+EXCEPTIONAL CASE:
+5 3 1
+sum = 4
+maxi = 5
+here maxi>sum but the answer is still 1 because cout<<maxi-sum which gives 5-4=1
+DOUBT CASE:
+5 3 0
+In this case it is not possible to bring all to zero playing alternatively
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1649/problem/C
+Vector used inside a map
+Exactly the same code but different template
+a,b,c,d
+to find |b-a|+|c-a|+|d-a| we did
+sort
+sum - (v.size()-1-i)*a
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1715/problem/A
+the equation is 2(smaller-1)+(larger-1)+1
+              =>2*smaller(n)-2+larger
+2(smaller-1) is because both stanley and megan mandatorily have to move that number of steps
+(larger-1) is the portal that megan is creating for stanley
++1 is the step required by Stanley to go through that portal
+Explanation for the DOUBT that I had is:
+we always multiply the smaller with 2 because we want to minimize
+also if you don't use smaller*2 it is giving a wrong answer in testcase 1 5
+DOUBT:
+for the testcase 1 5 the answer should be 4??Because of 5-1
+
+
+
+
+
+
+https://codeforces.com/contest/1742/problem/E
+KEEP IT UP
+for passing the first 4 testcases on a 1200 rated question but TLE.
+
+
+
+
+
+https://codeforces.com/contest/1732/problem/D1
+Weird shit happening here:
+If I declare the map and set before taking the testcase count input it gives the correct output otherwise it gives the wrong output
+#DOUBT
+
+
+
+
+
+https://codeforces.com/contest/1732/problem/B
+^mod
+#REDO
+
+
+
+
+
+
+https://codeforces.com/contest/677/problem/B
+
+
+
+
+
+
+
+https://codeforces.com/contest/1754/problem/B
+Making 2 arrays one with the smaller numbers the other made using the other half of the bigger numbers.
+output one large, one small number
+if number of numbers(n) is odd:
+using if(i<b.size);
+also while pushing back in tha arrays b:
+if(i<(n+1)/2)
+a.pb(i+1)
+else
+b.pb(i+1);
+
+
+
+
+
+
+https://codeforces.com/contest/1754/problem/C1
+We have to make segments such as the sum of the differences of all segments is 0;
+So if n is odd its just not possible because there will always be one integer remaining;
+In the question it is given that a segment can consist of a single integer also (li<=ri)
+so what we do is the numbers whose sum we get as non-zero we put them in one segment like {1,1}{-1,-1} so what will happen is within the segment their difference will be 0 i.e. {1-1}+{-1-(-1)}
+and for the numbers whose sum is already 0 we will consider that single integer as one segment i.e. 1,-1 so {1}+{-1} as we can take (li<=ri).
+
+
+
+
+
+
+
+https://codeforces.com/contest/1744/problem/B
+my attempt TLE(bruteforce)
+Optimised approach:
+To begin with
+take sum of all sum+=a[i];
+count number of odd and even n_odd++, n_even++;
+if the number that is supposed to be added(queries) to all the integers of the vector is odd (if query%2==1) then
+if type is 0:
+sum will be sum+=query*n_even
+then all even numbers will be turned odd for the next input of >>type>>queries.
+if type is 1:
+sum will be sum+=query*n_odd
+Then all odd numbers will become even for the next input of >>type>>queries.
+#CONCEPT
+adding odd number to an even number makes that number odd.
+adding odd number to an odd number makes that number even.
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1744/problem/C
+We append the string s after string s so the our string becomes like "rggryrggry", we do this because after the last 'r' if it does not find any 'g' it starts from 0th index again until it finds a 'g'
+now we store the index of all the occurences of green light 'g' in a vector.
+We want the green signal 'g' to be in front of our 'r', to make that sure we use j to iterate over the vector to find a 'g' whose index is greater than 'r' we use while(pos_g[j]<i) j++;pos_g[j] gives us the index of the 'g'.
+As soon as we find our first 'g' we caculate the distance of that 'g' from 'r' (pos_g[j]-i)
+And our objective is to find the maximum distance between 'r' and the first occurence of 'g' on the right of that 'r'.
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1744/problem/A
+using a map of vectors to store the indices of that integer given in the input array.
+using it.second to iterate over the array in the map checking if same char of the string is denoted by same integer in the array.
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1736/problem/A
+I didn't catch that if the number of 1's is same then changing the order requires only 1 move so my code was incorrect.
+count the number of 1's.
+If number of 1's is same check if the the vectors itself are identical or not
+If they are identical no moves are required but if they are not 1 move will be required to change the order.
+Calculate the difference of 1's
+#DOUBT Testcase:
+1 0 1 0
+0 1 0 1
+
+
+
+
+
+
+
+https://codeforces.com/contest/1736/problem/B
+Using continue instead of return gives 5 outputs in the first testcase.
+FORMULA USED:
+a*b/__gcd(a,b) = lcm(a,b);
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1721/problem/C
+Using list<int> to calculate d_min=list.back()-a[i]   and    d_max=list.front()-a[i]  to use funtions .back() and .front()
+the bptr will only move behind if b[bptr]>=a[i] so that the values of d_min and d_max don't come negative
+We are clearing the list(larger_than_a.clear()) if we the preceding b element is smaller than the current a element and the count of terms visited (acnt) and count of b terms visited (bcnt) is same
+NOTE THAT if all the values in the b vector are greater than the existing a then the bptr will reach the 0th index and will sit there throughout the execution
+n-(btr+1) to find the elements visited in the b vector, bptr+1 because bptr has gone a step back after the iteration b--;
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1717/problem/B
+As we are using 0 based indexing in our code we do r--, c-- the coordinates where 'X' is already present.
+vector<vector<char>> v(n, vector<char>(n, '.')); 2d char vector of '.''s
+v[r][c] = v[horizontal number][vertical number]
+x=r y=c because we are starting from the 'X' that is already present in the matrix as given in the question
+while(x>=0){//while we are moving upwards
+ x--;
+ y++;
+ because we are placing in a pattern  .........x
+                                        ......x.
+ if(y>n)
+ {
+    gave wrong output it should rather be if(y>=n) maybe beacause we are using 0 based indexing so =n is also going out of the matrix
+ }
+}
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1747/problem/B
+if(n%2):
+"BANBANBAN"
+cout<<(n/2)+1<<endl; because n/2 moves until the middle "BAN" is left and 1 extra move to swap letters for that middle "BAN"
+i=1 pointer on the first'B'
+j=3*n pointer on the last 'N'
+also:
+cout<<2<<" "<<j; for swapping that middle BAN's 'N' and the first BAN's 'A'.
+if(n%2==0):
+in the even case no +1 required in (n/2)+1
+and also no cout<<2<<" "<<j required in the end.
+for(i=0; i<k/2)
+i<k/2 because k/2 is the number of swaps needed
+and i+=3, j-=3
+for i to jump from one 'B' to the next 'B' AND
+for j to jump from one 'N' to the previous 'N'.
+
+
+
+
+
+
+
+https://codeforces.com/contest/1747/problem/C
+using *min_element(a,a+n);
+alice takes the first chance to reduce the first element by 1 and swap the first element with any other element
+Then Bob has to reduce the first element by 1 and he swaps it with any other element.
+The person who sees a 0 in the first place loses
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1237/A
+An even number is kept as it is
+Half of the odd numbers are decreased by 1 and the rest half are increased by 1
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1735/problem/B
+Even the other code that I submitted could be right as in that still if you cut by 2*a[0] the pieces will always be <= 2*a[0]
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1658/problem/B
+To get an even number we always have to do even*even or odd*even
+MATH:
+COMBINATORICS:
+if n=6
+1 2 3 4 5 6
+1 is odd, for 1 the numbers that can be used are the even ones that are 2,4,6
+The explanation is quite long and it is recommended to watch the video
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1733/problem/C
+Accepted even if testcases don't match
+if first element is odd we output the indices of all the odd elements 2 at a time
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1748/problem/B
+Finding the longest substring where the frequency of numbers does not exceed the number of distinct numbers
+for(int i=0; i<n; i++)
+       {
+          vector<int> freq(10);
+          int distinct = 0, maxi = INT_MIN;
+          for(int j=i; j<n; j++)
+          {
+             if(j-i>=101) break;
+
+             int c = s[j] - '0';
+             if(freq[c]==0) distinct++;
+
+             maxi = max(maxi, ++freq[c]);
+
+             if(maxi<=distinct)
+               ans++;
+          }
+          cout<<ans<<endl;
+NOTE:
+++freq[c], Here we are incrementing first and then comparing it with maxi aka pre-increment
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1717/problem/C
+Using b[(i+1)%n] here we used %n because after the (n-1)th index the 0th index is considered so the array is called cyclic
+It was really hard to grasp that the question meant that a[i+1] has to be >= a[i]-1
+and the maximum value that a[i+1] can get is b[i+1] therefore
+b[i+1] has to be >= b[i]-1
+According to the Newton school IIT-KGP youtube guy the questions means that a[i] can be <= a[i]+1 which gave a correct answer BUT
+In the question it is given a[i] can be <= a[i+1]
+#DOUBT
+
+
+
+
+
+
+https://codeforces.com/contest/1719/problem/B
+if k is odd then we can make pairs like odd even
+                                        odd even
+                                        odd even
+because we can add our odd k to the odd number and make it even as odd+odd=even
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/136/A
+Questions like these where the output is nothing, all that matters is understanding the questions well
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1760/problem/C
+Key Takeaway:
+Don't declare the size of the vector in which you are going to push_back
+Otherwise output will be all 0's 0 0 0 0
+                                 0 0 0
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1760/problem/D
+valley question
+#REDO
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1748/problem/C
+Prefix sum
+
+
+
+
+
+https://codeforces.com/problemset/problem/1765/K
+#REDO
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1764/C
+"maxi = max(maxi, (i+1)*(n-1-i))"
+This statement means that we are finding the maximum number of attachments of 3 numbers that can be made such that it is neither in ascending order nor in descending order
+i.e. a<b>c if a,b,c are the 3 numbers
+we are partitioning on the basis of distinct numbers so "111223" we will partition like "111" and "223"
+(i + 1) -> number of element till index i
+(n-1-i) -> number of element after index i
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1761/C
+Recursive DFS
+
+
+
+
+
+
+
+https://codeforces.com/contest/1749/problem/C
+Kuch bhi nahi samajh mein aaya code mein
+
+
+
+
+
+
+https://codeforces.com/contest/1720/problem/C
+dividing the matrix into smaller 2x2 matrices:
+if there are atleast 2 zeroes, number of 1's is the number of operations
+if there is atleast 1 zero, x-1 operations(where x is the number of 1's)
+0 zeroes,x-2 operations
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1669/G
+for(int j=0; j<m; j++)
+          {
+             for(int i=n-2; i>=0; i--)
+             {
+                if(a[i][j]=='*')
+                {
+                   int k=i+1;
+                   while(k<n && a[k][j]=='.')
+                        k++;
+
+                   swap(a[k-1][j], a[i][j]);
+                }
+             }
+          }
+KEY to underestand was to start checking from down i.e. from i=n-2
+because if a stone falls down it will allow the stone above it to fall right above it.
+
+
+
+
+
+
+
+https://codeforces.com/contest/1669/problem/D
+          for(int i=0; i<n; i++)
+          {
+              while(i<n && s[i]=='W')
+                   i++;
+              if(i==n)
+                break;
+              int red = 0;
+              int blue = 0;
+              while(i<n && s[i]!='W')
+              {
+                 if(s[i]=='R')
+                  red++;
+                 else
+                   blue++;
+
+                 i++;
+              }
+              if(red==0 || blue==0)
+              {
+                 flag = false;
+                 break;
+              }
+          }
+
+
+
+
+
+https://codeforces.com/contest/1470/problem/A
+          int n,m;
+          cin>>n>>m;
+
+          int k[n+1];
+          for(int i=1; i<=n; i++)
+            cin>>k[i];
+          int c[m+1];
+          for(int i=1; i<=m; i++)
+             cin>>c[i];
+
+          int gift[n];
+          for(int i=1; i<=n; i++)
+          {
+             gift[i] = c[k[i]];
+          }
+
+          sort(gift+1, gift+n+1);
+
+          int amount = 0;
+          int j = 1;
+          for(int i=n; i>0; i--)
+          {
+             if(c[j]<gift[i])
+             {
+                 amount+=c[j];
+                 j++;               //only if c[j] is smaller than gift[i] we are incrementing j++, otherwise only i decrements i--
+             }
+             else
+              amount+=gift[i];
+          }
+          cout<<amount<<endl;
+
+
+
+
+
+https://codeforces.com/contest/1669/problem/F
+#TWO POINTERS
+          while(i<=j)
+                    {
+                       if(right==left && right>0)
+                       {
+                          ans=n-j+i-1;            we update the ans whenever left==right and they are not zero
+                       }
+                       if(left==right)
+                       {
+                          left+=a[i];
+                          i++;
+                       }
+                       else if(left>right)
+                       {
+                          right+=a[j];
+                          j--;
+                       }
+                       else
+                       {
+                          left+=a[i];
+                          i++;
+                       }
+                    }
+                    if(left==right) ans = n;
+                    cout<<ans<<endl;
+
+
+
+https://codeforces.com/contest/1702/problem/E
+#REDO
+#THEORY, proof of Bipartite Graphs
+
+
+
+
+
+
+https://codeforces.com/contest/1693/problem/A
+for(int i=0; i<n-1; i++)          we only iterate till n-1 because the last element has to be zero because we got to come back from it which will decrease its value by -1
+          {
+             if(v[i]<=0)          if the prefix sum ever goes to 0 or below we break and now for it to be valid all numbers in front of it have to be zero
+                break;            The observation says that the number of negatives in the vector must be equal to the number of positives in the first element
+             v[i+1]+=v[i];
+             v[i]=0;
+          }
+          bool all_zero = true;
+          for(int i=0; i<n; i++)
+          {
+               if(v[i] != 0)
+               {
+                  all_zero = false;
+               }
+          }
+          if(all_zero)
+            yes;
+          else
+            no;
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1692/G
+int n,k;
+          cin>>n>>k;
+
+          int a[n];
+
+          for(int i=0; i<n; i++)
+          {
+             cin>>a[i];
+          }
+
+          int ans = 0;
+          int cnt = 1;
+
+          for(int i=1; i<n; i++)
+          {
+              if(a[i-1]<2LL*a[i])
+                cnt++;
+              else
+                cnt=1;
+
+              if(cnt>k)
+                ans++;
+          }
+          cout<<ans<<endl;
+
+
+
+
+
+
+https://codeforces.com/group/MWSDmqGsZm/contest/223339/problem/I
+RECURSION
+
+
+
+
+
+
+https://codeforces.com/group/MWSDmqGsZm/contest/223339/problem/S
+RECURSION
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/451/B
+REVERSE(a+start, a+end);
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/816/B
+#STRANGE OBSERVATION
+prefix sum has the ability to count the frequency in a range
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1195/C
+int h[2][N];
+int dp[2][N];   these must be initialized outside int main()
+dynamic programming
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1355/B
+total is the number of GROUPS that are formed
+left is the number of PEOPLE that are left
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/230/B
+sieve for getting all prime numbers till 10^6
+set<int> unique_squares;
+ bool a[1000005];
+
+  void sieve()
+  {
+     for(int i=2; i<1000000; i++)
+     {
+         if(!a[i])
+         {
+             unique_squares.insert(i*i);
+             for(int j=i*i; j<1000000; j+=i)
+             {
+                a[j] = true;
+             }
+         }
+     }
+  }
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1360/D
+for(int i=1; i*i<=n; i++)
+       {
+           if(n%i==0)
+           {
+               if(i<=k)
+               {
+                  ans = min(ans, n/i);
+               }
+               if(n/i<=k)
+               {
+                  ans = min(ans, i);
+               }
+           }
+       }
+if n=8;
+   k=10
+   i=1
+the trick here was that you can buy 1 package of 8 shovels and 8 packages of 1 shovel
+and we needed minimum package thats why we check if n/i<=k which means
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1294/C
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1389/A
+void solve ()
+{
+    int a,b;
+    cin>>a>>b;
+
+    if(2LL*a<=b)
+    {
+        cout<<a<<" "<<2LL*a<<"\n";
+    }
+    else
+    {
+       cout<<-1<<" "<<-1<<"\n";
+    }
+}
+
+
+
+
+
+
+https://codeforces.com/contest/1617/problem/C
+
+       int n;
+       cin>>n;
+
+       int a[n];
+
+       map<int,int> mpp;
+
+       vector<int>notinpermutation;
+
+       for(int i=0; i<n; i++)
+       {
+          cin>>a[i];
+          mpp[a[i]]++;
+          if(mpp[a[i]]>1 && a[i]<=n)
+          {
+              notinpermutation.pb(a[i]);
+          }
+          else if(a[i]>n)
+          {
+              notinpermutation.pb(a[i]);
+          }
+       }
+
+       sort(notinpermutation.begin(), notinpermutation.end());
+       int number_of_permutations = notinpermutation.size();
+       reverse(notinpermutation.begin(), notinpermutation.end());
+       for(int i=0; i<n; i++)
+       {
+          if(mpp[i+1]==0)
+          {
+             int z = notinpermutation.back();
+             notinpermutation.pop_back();
+             int y = i+1;
+             if(z%(z-y)==y)
+               continue;
+             else
+             {
+                cout<<"-1\n";
+                return;
+             }
+          }
+       }
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1613/C
+#BINARYSEARCH
+#BESTPROBLEM
+https://www.youtube.com/watch?v=LQrBi4D3WoU
+Whenever the answer lies in a range that is sorted, use binary search
+
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1605/problem/C
+vector<int> v;
+
+      bool two=false, three=false, four=false, seven=false;
+
+      for(int i=0; i<n; i++)
+      {
+         if(s[i]=='a')
+         {
+            v.pb(i);
+         }
+      }
+      if(v.size()==0 || v.size()==1)
+      {
+         cout<<"-1\n";
+         return;
+      }
+      for(int i=0; i<v.size()-1; i++)
+      {
+         if(v[i+1]-v[i]==1)
+         {
+            two = true;
+            break;
+         }
+         else if(v[i+1]-v[i]==2)
+         {
+            three = true;
+         }
+         else if(v[i+1]-v[i]==3)
+         {
+            if(s[v[i]+1] != s[v[i]+2])
+            {
+                four = true;
+            }
+            else
+            {
+               if(i+1<v.size()-1)
+               {
+                  if(v[i+2]-v[i+1]==3)
+                  {
+                     if(s[v[i+1]+1]==s[v[i+1]+2])
+                     {
+                        if(s[v[i+1]+1] != s[v[i]+1])
+                        {
+                           seven = true;
+                        }
+                     }
+                  }
+               }
+            }
+         }
+      }
+      if(two)
+        cout<<"2\n";
+      else if(three)
+        cout<<"3\n";
+      else if(four)
+        cout<<"4\n";
+      else if(seven)
+        cout<<"7\n";
+      else
+        cout<<"-1\n";
+
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1579/D
+int n;
+      cin>>n;
+      priority_queue<pair<int,int>> pq;
+      for(int i=1; i<=n; i++)
+      {
+         int x;
+         cin>>x;
+         if(x>0){
+         pq.push({x,i});
+        }
+      }
+      vector<pair<int,int>>ans;
+      while(pq.size()>1)
+      {
+         auto f =pq.top();
+         pq.pop();
+         auto s = pq.top();
+         pq.pop();
+         ans.pb({f.second, s.second});
+         f.first--;
+         s.first--;
+         if(f.first) pq.push(f);
+         if(s.first) pq.push(s);
+      }
+      cout<<ans.size()<<endl;
+      for(auto x:ans)
+      {
+         cout<<x.first<<" "<<x.second<<endl;
+      }
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1535/C
+string s;
+      cin>>s;
+      int n = s.size();
+      int zero = 0, one = 0;
+      int ans = 0;
+
+      for(int i=0; i<n; i++)
+      {
+         if(s[i]=='0')
+         {
+            zero++;
+            one=0;
+            ans+=zero;
+         }
+         else if(s[i]=='1')
+         {
+            one++;
+            zero=0;
+            ans+=one;
+         }
+         else
+         {
+            zero++;
+            one++;
+            ans+=max(zero,one);
+         }
+         swap(zero,one);
+      }
+      cout<<ans<<endl;
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1451/C
+int n,k;
+       cin>>n>>k;
+       string a,b;
+       cin>>a>>b;
+       vector<int>cnt1(26),cnt2(26);
+       for(auto x:a)
+       {
+          cnt1[x-'a']++;
+       }
+       for(auto x:b)
+       {
+          cnt2[x-'a']++;
+       }
+       for(int i=0; i<26; i++)
+       {
+          if(cnt1[i]<cnt2[i])
+          {
+              no;
+              return;
+          }
+          cnt1[i] -= cnt2[i];
+          if(cnt1[i]%k>0 || (i==25&&cnt1[i]>0))
+          {
+              no;
+              return;
+          }
+          if(i<=24)
+            cnt1[i+1]+=cnt1[i];
+       }
+       yes;
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1443/C
+bool comp(pair<int,int> a,pair<int,int> b)         to sort w.r.t first number in the pair
+{
+    if(a.first != b.first) return a.first > b.first;
+    else
+    return a.second > b.second;
+}
+
+void solve ()
+   {
+       int n;
+       cin>>n;
+       int a[n];
+       int b[n];
+       for(int i=0; i<n; i++)
+          cin>>a[i];
+
+       for(int i=0; i<n; i++)
+         cin>>b[i];
+
+       vector<pair<int,int>> v;
+
+       for(int i=0; i<n; i++)
+       {
+          v.pb(make_pair(a[i],b[i]));
+       }
+
+       sort(v.begin(), v.end(), comp);
+       int sum = 0;
+       for(int i=0; i<v.size(); i++)
+       {
+          if(sum >= v[i].first)
+            continue;
+          if(v[i].first <= sum+v[i].second)
+            sum=v[i].first;
+          else
+            sum+=v[i].second;
+       }
+       cout<<sum<<endl;
+   }
+
+
+
+
+
+
+   int a,b,k;
+          cin>>a>>b>>k;
+          int boys[k], cnt1[a+1] = {0};
+          int girls[k], cnt2[b+1] = {0};
+
+          for(int i=0; i<k; i++)
+          {
+             cin>>boys[i];
+             cnt1[boys[i]]++;
+          }
+          for(int i=0; i<k; i++)
+          {
+             cin>>girls[i];
+             cnt2[girls[i]]++;
+          }
+          int cnt = 0;
+          for(int i=0; i<k; i++)
+          {
+             cnt+=k-(cnt1[boys[i]] + cnt2[girls[i]]) + 1;
+          }
+          cout<<cnt/2<<endl;
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1644/C
+          int n,x;
+                 cin>>n>>x;
+                 int a[n];
+                 for(int i=0; i<n; i++)
+                    cin>>a[i];
+
+                 vector<int> ans1(n+1, INT_MIN);
+                 ans1[0] = 0;
+
+                 for(int i=0; i<n; i++)
+                 {
+                     int temp = 0;
+                     for(int j=i; j<n; j++)
+                     {
+                        temp+=a[j];
+                        ans1[j-i+1] = max(ans1[j-i+1], temp);
+                     }
+                 }
+                 for(int i=0; i<=n; i++)
+                 {
+                    int ans = 0;
+                    for(int j=0; j<=n; j++)
+                    {
+                        ans=max(ans, min(j,i)*x + ans1[j]);
+                    }
+                    cout<<ans<<" ";
+                 }
+                 cout<<endl;
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1548/A
+easy code
+Game Of Thrones question
+          int n,m;
+          cin>>n>>m;
+
+          int friends[n] = {0};
+          int loyal=n;
+
+/*          for(int i=0; i<m; i++)
+          {
+              int x,y;
+              cin>>x>>y;
+
+              if(x>y)
+                swap(x,y);
+
+              x--; y--;         //for 0 based indexing
+              if(friends[x]==0)
+              {
+                loyal--;
+              }
+
+              friends[x]++;             making the Friendship Graph */
+          }
+          int t;
+          cin>>t;
+          while(t--)
+          {
+              int query;
+              cin>>query;
+
+              if(query==1)
+              {
+                  int x,y;
+                  cin>>x>>y;
+                  if(x>y)
+                    swap(x,y);
+
+                  x--; y--;
+                  if(friends[x]==0)
+                  {
+                    loyal--;
+                  }
+
+                  friends[x]++;
+              }
+              if(query==2)
+              {
+                 int x,y;
+                 cin>>x>>y;
+                 if(x>y)
+                  swap(x,y);
+
+                  x--; y--;
+
+                  friends[x]--;
+                  if(friends[x]==0)
+                  {
+                     loyal++;
+                  }
+              }
+              if(query==3)
+              {
+                  cout<<loyal<<endl;
+              }
+          }
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1540/A
+https://www.youtube.com/watch?v=4D65uGUMtPk
+          int n;
+          cin>>n;
+          int a[n];
+          for(int i=0; i<n; i++)
+          {
+              cin>>a[i];
+          }
+
+          sort(a,a+n);
+          int sum=a[n-1];
+
+          int neg[n];
+          neg[0]=0;
+          for(int i=1; i<n; i++)
+          {
+             neg[i]=neg[i-1] + i*(a[i]-a[i-1]);//     (A+B) + 3*(X)     explained here https://www.youtube.com/watch?v=4D65uGUMtPk
+             sum-=neg[i];
+          }
+          cout<<sum<<endl;
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1517/C
+#include<bits/stdc++.h>
+using namespace std;
+#define int long long
+#define endl "\n"
+const int mod=1000000007;
+mt19937_64 rng(std::chrono::steady_clock::now().time_since_epoch().count());
+int arr[501][501];
+void solve(){
+      int n;
+      cin>>n;
+
+      int rem[n+1];
+      for(int i=0;i<n;i++){
+            rem[i+1]=i;
+            cin>>arr[i][i];
+      }
+
+      bool vis[n][n];
+      memset(vis,0,sizeof(vis));
+      // for(int i=0;i<n;i++){
+      map<int,pair<int,int>>mp;
+      for(int i=0;i<n;i++){
+            vis[i][i]=1;
+            mp[i]={i,i};
+      }
+      bool flag=1;
+      int cnt=0;
+      set<int>st;
+      for(int i=0;i<n;i++){
+            st.insert(i);
+      }
+      while(!st.empty()){
+            // cnt++;
+            // if(cnt==1000){
+            //       cout<<-1;
+            //       return;
+            // }
+            flag=0;
+            vector<int>temp;
+            for(auto i:st){
+                  int tot=0;
+                  int x=mp[i].first;
+                  int y=mp[i].second;
+                  if(rem[arr[x][y]]==0){
+                        temp.push_back(i);
+                        continue;
+                  }
+                  if(x-1>=0 and (x-1)>=y and !vis[x-1][y]){
+                        tot++;
+                  }
+                  if(x+1<n and (x+1)>=y and !vis[x+1][y]){
+                        tot++;
+                  }
+                  if(y-1>=0 and (y-1)<=x and !vis[x][y-1]){
+                        tot++;
+                  }
+                  if(y+1<n and (y+1)<=x and !vis[x][y+1]){
+                        tot++;
+                  }
+                  if(tot==1){
+                        flag=1;
+                        if(x-1>=0 and (x-1)>=y and !vis[x-1][y]){
+                              mp[i]={x-1,y};
+                        }
+                        if(x+1<n and (x+1)>=y and !vis[x+1][y]){
+                              mp[i]={x+1,y};
+                        }
+                        if(y-1>=0 and (y-1)<=x and !vis[x][y-1]){
+                              mp[i]={x,y-1};
+                        }
+                        if(y+1<n and (y+1)<=x and !vis[x][y+1]){
+                              mp[i]={x,y+1};
+                        }
+                        rem[arr[i][i]]-=1;
+                        vis[mp[i].first][mp[i].second]=1;
+                        arr[mp[i].first][mp[i].second]=arr[i][i];
+                  }
+            }
+            for(auto z:temp){
+                  st.erase(z);
+            }
+      }
+      for(int i=0;i<n;i++){
+            for(int b=0;b<=i;b++){
+                  cout<<arr[i][b]<<" ";
+            }
+            cout<<endl;
+      }
+
+      // cout<<endl;
+}
+
+int32_t main(){
+      ios_base::sync_with_stdio(0);
+      cin.tie(0);
+      // srand(time(0));
+      int t;
+      // cin>>t;
+      t=1;
+      while(t--){
+            solve();
+      }
+      return 0;
+}
+It was accepted before but now it shows TLE
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1553/C
+int end(int w1, int w2, int rem1, int rem2)
+{
+    if(w1>w2+rem2 || w2>w1+rem1)
+    {
+        return 1;
+    }
+    else
+    {
+        return 0;
+    }
+}
+
+int m1(string s)
+{
+    int w1=0,w2=0,rem1=5,rem2=5;
+    int ans=0;
+    for (int i = 0; i < 10; i++)
+    {
+        if(end(w1,w2,rem1,rem2))
+        {
+            return ans;
+        }
+        if(s[i]=='1' || s[i]=='?')
+        {
+            w1++;
+            rem1--;
+        }
+        else if (s[i] == '0')
+        {
+            rem1--;
+        }
+        ans++;
+        i++;
+        if (end(w1, w2, rem1, rem2))
+        {
+            return ans;
+        }
+
+        if (s[i] == '1')
+        {
+            w2++;
+            rem2--;
+        }
+        else if (s[i] == '0' || s[i] == '?')
+        {
+            rem2--;
+        }
+        ans++;
+    }
+
+    return 10;
+
+}
+
+int m2(string s)
+{
+    int w1 = 0, w2 = 0, rem1 = 5, rem2 = 5;
+    int ans = 0;
+    for (int i = 0; i < 10; i++)
+    {
+        if (end(w1, w2, rem1, rem2))
+        {
+            return ans;
+        }
+        //a
+        if (s[i] == '1' )
+        {
+            w1++;
+            rem1--;
+        }
+        else if (s[i] == '0' || s[i] == '?')
+        {
+            rem1--;
+        }
+        ans++;
+
+        if (end(w1, w2, rem1, rem2))
+        {
+            return ans;
+        }
+
+        i++;
+
+        if (s[i] == '1' || s[i] == '?')
+        {
+            w2++;
+            rem2--;
+        }
+        else if (s[i] == '0' )
+        {
+            rem2--;
+        }
+        ans++;
+    }
+
+    return 10;
+}
+
+int main()
+{
+    ios_base::sync_with_stdio(false);
+    cin.tie(NULL);
+    int t;
+    cin >> t;
+    for (int tt = 0; tt < t; tt++)
+    {
+        string s;
+        cin>>s;
+        cout<<min(m1(s),m2(s))<<"\n";
+
+    }
+    return 0;
+
+}
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1343/C
+        int n;
+        cin>>n;
+
+        int a[n+1];
+        for(int i=1; i<=n; i++)
+        {
+            cin>>a[i];
+        }
+
+        set<int> s;
+        int i=1;
+        int sum = 0;
+        while(i<=n)
+        {
+           if(a[i]<0)
+           {
+              while(a[i]<0 && i<=n)
+              {
+                  s.insert(a[i]);
+                  i++;
+              }
+           }
+           else
+           {
+              while(a[i]>0 && i<=n)
+              {
+                 s.insert(a[i]);
+                 i++;
+              }
+           }
+
+           auto it = s.rbegin();
+           sum+=*it;
+           s.clear();
+
+        }
+        cout<<sum<<endl;
+
+arranging the number in the order +,-,+,-
+and finding the max sum of the subsequence
+so in the "set" we store:
+1)The least negative number AND
+2)The maximum positive number
+ACTUALLY DIDN'T UNDERSTAND
+why
+auto it = s.rbegin();
+sum+=*it;
+s.clear();
+is coming inside the loop
+
+
+
+
+
+
+https://codeforces.com/contest/1324/problem/D
+#include <bits/stdc++.h>
+
+using namespace std;
+
+#define int long long
+#define pb push_back
+#define yes cout<<"YES"<<endl
+#define no cout<<"NO"<<endl
+
+
+
+
+void solve()
+{
+     int n;
+     cin>>n;
+
+     int a[n];
+     int b[n];
+
+     for(int i=0; i<n; i++)
+     {
+        cin>>a[i];
+     }
+     for(int i=0; i<n; i++)
+     {
+          cin>>b[i];
+     }
+
+     int c[n];
+
+
+     for(int i=0; i<n; i++)
+     {
+        c[i] = b[i]-a[i];
+     }
+     sort(c, c+n);
+     int ans = 0;
+     for(int i=0; i<n; i++)
+     {
+        auto it = lower_bound(c+i+1, c+n, -c[i]);
+        int id = it-c;
+        ans += id-(i+1);
+     }
+
+     cout<<ans<<endl;
+}
+
+int32_t main(){
+      ios_base::sync_with_stdio(0);
+      cin.tie(0);
+   //   int t;
+   //   cin>>t;
+   //   while(t--)
+      {
+        solve();
+      }
+      return 0;
+}
+lower_bound() to find the first number that is greater than or equal to -ci
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1375/C
+int n;
+     cin>>n;
+     stack<int> s;
+     for(int i=0; i<n; i++)
+     {
+          int x;
+          cin>>x;
+          if(s.empty())
+          {
+             s.push(x);
+          }
+          else
+          {
+              int val;
+              while(!s.empty() && s.top()<x)
+              {
+                 val = s.top();
+                 s.pop();
+              }
+              s.push(min(val,x));
+          }
+     }
+     if(s.size()==1)
+      yes;
+    else
+      no;
+using stack
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1350/problem/B
+map<int,vector<int>> mpp;
+     for(int i=1;i<=100003; i++)
+     {
+        for(int j=1; j<=sqrt(i); j++)
+        {
+           if(i%j==0)
+           {
+              if(i/j==j)
+              {
+              	 mpp[i].pb(j);
+              }
+             else
+             {
+                mpp[i].pb(j);
+                mpp[i].pb(i/j);
+             }
+           }
+        }
+     }
+
+    int t;
+    cin>>t;
+    while(t--)
+    {
+       int n;
+       cin>>n;
+
+       int a[n+1];
+       for(int i=1; i<=n; i++)
+          cin>>a[i];
+
+       int ans[n+1];
+       for(int i=0; i<=n; i++)
+       {
+          ans[i]=1;
+       }
+
+       for(int i=1; i<=n; i++)
+       {
+          for(auto x:mpp[i])
+          {
+          	 if(a[x]<a[i])
+          	 {
+                ans[i] = max(ans[i], ans[x]+1);
+             }
+          }
+       }
+       int final_ans = INT_MIN;
+       for(int i=1; i<=n; i++)
+       {
+       	  if(final_ans<ans[i])
+       	  	final_ans = ans[i];
+       }
+
+    cout<<final_ans<<endl;
+    }
+map of vectors of squares
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1418/B
+If x==0
+push a[i] in b vector
+assign a[i] as infinite
+now sort b in descending order sort(b.begin(), b.end(), greater<int>());
+now assign those values in a as the first values in b(which are the greatest)
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/732/problem/D
+bool of(int mid)
+{
+    set<int> s;
+    int days_left = 0;
+    for(int i=mid; i>=0; i--)
+    {
+        if(d[i]!=0 && s.count(d[i])==0)
+        {
+            s.insert(d[i]);
+            days_left+=a[d[i]-1];
+        }
+        else
+          days_left = max(0LL, --days_left);
+    }
+    if(days_left==0 && s.size()==m)
+      return true;
+    else
+      return false;
+}
+
+
+void solve()
+{
+
+
+        cin>>n>>m;
+        for(int i=0; i<n; i++)
+        {
+           cin>>d[i];
+        }
+        for(int i=0; i<m; i++)
+        {
+            cin>>a[i];
+        }
+        int start = 0;
+        int end = n-1;
+        while(start<end)
+        {
+           int mid=start+(end-1)/2;
+           if(of(mid))
+           {
+              end = mid;
+           }
+           else
+            start=mid+1;
+        }
+        if(of(start))
+          cout<<start+1;
+        else
+          cout<<-1;
+}
+This code gives TLE
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/978/C
+int n,m;
+     cin>>n>>m;
+     int a[n];
+     for(int i=0; i<n; i++)
+     {
+        cin>>a[i];
+     }
+     for(int i=1; i<n; i++)
+     {
+        a[i]+=a[i-1];
+     }
+     set<pair<int,int>>s;
+     s.insert({0,0});
+     for(int i=0; i<n; i++)
+     {
+        s.insert({a[i],i+1});
+     }
+     for(int i=0; i<m; i++)
+     {
+        int x;
+        cin>>x;
+        auto it = s.lower_bound({x,0});
+        cout<<(*it).second<<" ";
+        --it;
+        cout<<  x-(*it).first<<endl;
+     }
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1366/B
+int n,x,m;
+     cin>>n>>x>>m;
+     int a=x;
+     int b=x;
+     int l,r;
+     for(int i=0; i<m; i++)
+     {
+        cin>>l>>r;
+        if(l<=a && r>=a)
+          a=l;
+        if(l<=b && r>=b)
+          b=r;
+     }
+    cout<<b-a+1<<endl;
+#2 POINTERS
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1076/A
+for(int i=0; i<n-1; i++)
+     {
+          if(s[i]>s[i+1])
+          {
+             s[i] = '1';
+             flag = true;
+             break;
+          }
+     }
+     if(!flag)
+      s[n-1] = '1';
+
+     for(auto x:s)
+     {
+        if(x!='1')
+          cout<<x;
+     }
+#LEXICOGRAPHICAL STRINGS
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1251/problem/B
+int one = 0, zero=0, len=0;
+
+     for(int i=0; i<n; i++)
+     {
+        string s;
+        cin>>s;
+
+        for(auto x:s)
+        {
+           if(x=='1')
+            one++;
+          else
+            zero++;
+        }
+         len += (s.size()/2);
+     }
+
+    if(one/2+zero/2 >= len)
+     cout<<n<<endl;
+    else
+       cout<<n-1<<endl;
+#PALINDROMES
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/946/C
+string s;
+     cin>>s;
+
+     bool flag = false;
+     int j = 0;
+     for(int i=0; i<s.length(); i++)
+     {
+        if(s[i]-'a'<=j)
+        {
+            s[i] = char(j + 'a');
+            j++;
+        }
+        if(j==26)
+        {
+            flag = true;
+            break;
+        }
+     }
+     if(flag)
+      cout<<s<<endl;
+     else
+     cout<<-1<<endl;
+#ASCII
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/628/C
+for(int i=0; i<n; i++)
+      {
+          if(s[i]-k >= 'a')
+          {
+             s[i]-=k;
+             k=0;
+          }
+          else if(s[i] + k <= 'z')
+          {
+              s[i]+=k;
+              k=0;
+          }
+          else
+          {
+             int a=abs(s[i] - 'a'), b=abs(s[i] - 'z');
+             if(a<=b)
+             {
+                k-=b;
+                s[i]='z';
+             }
+             else
+             {
+                k-=a;
+                s[i]='a';
+             }
+          }
+      }
+      if(k)
+        cout<<-1;
+      else
+        cout<<s;
+#REDO
+#ascii
+#STRINGS
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1425/A
+int n;
+      cin>>n;
+
+      int a=0, b=0, p=0;
+
+       while(n>0)
+       {
+          if((n==4 || n%4) && n%2==0)
+          {
+             if(p==0)
+              a+=n/2;
+             else
+              b+=n/2;
+
+            n /= 2;
+          }
+          else
+          {
+              if(p==0)
+                a++;
+              else
+                b++;
+
+               n--;
+          }
+          p = !p;
+       }
+       cout<<a;
+       cout<<endl;
+#MATH
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1291/B
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/670/D1
+int n,k;
+      cin>>n>>k;
+      vector<int> req(n);
+      vector<int> have(n);
+      for(int i=0; i<n; i++)
+        cin>>req[i];
+      for(int i=0; i<n; i++)
+        cin>>have[i];
+
+      int low=0, high=2001;
+      while(low<=high)
+      {
+          int mid = low+(high-low)/2;
+          int zer=0, cost=0;
+          for(int i=0; i<n; i++)
+            cost+=max(req[i]*mid-have[i], zer);
+          if(cost<=k)
+            low=mid+1;
+          else
+            high=mid-1;
+      }
+
+     cout<<high;
+#BINARYSEARCH
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1705/problem/C
+int n,c,q;
+     cin>>n>>c>>q;
+     string s;
+     cin>>s;
+
+     vector<pair<int,int>> len(c), query(c);
+
+
+     int curr = n;
+
+     for(int i=0; i<c; i++)
+     {
+        cin>>query[i].first>>query[i].second;
+        int qlen = query[i].second - query[i].first + 1;
+
+        len[i].first = curr+1, len[i].second = curr+qlen;
+        curr = len[i].second;
+     }
+     while(q--)
+     {
+         int idx;
+         cin>>idx;
+
+         for(int i=c-1; i>=0; i--)
+         {
+             if(idx>=len[i].first && idx<=len[i].second)
+             {
+                int gap = idx-len[i].first;
+                idx = query[i].first+gap;
+             }
+         }
+       cout<<s[idx-1]<<endl;
+     }
+#HARD
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1628/A
+vector<int> a(n);
+      vector<int> last(n+1, -1);
+      for(int i=0;i<n; i++)
+      {
+          cin>>a[i];
+          last[a[i]]=i;
+      }
+      vector<int> ans;
+      vector<int> mexv(n+1, false);
+      set<int> unique;
+      int mex = 0;
+      for(int i=0; i<n; i++)
+      {
+         mexv[a[i]]=true;
+         unique.insert(a[i]);
+         while(mexv[mex]==true)
+          mex++;
+        if(last[mex]<=i)
+        {
+           ans.pb(mex);
+           for(auto j:unique) mexv[j]=false;
+           mex=0;
+           unique.clear();
+        }
+      }
+      if(unique.size()) ans.pb(mex);
+      cout<<ans.size()<<endl;
+      for(auto x:ans)
+        cout<<x<<" ";
+MEX
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1276/A
+string s;
+      cin>>s;
+
+      vector<int> ans;
+
+      for(string t: {"twone", "two", "one"})
+      {
+          int pos = 0;
+
+          while((pos=s.find(t,pos)) != std::string::npos)
+          {
+             s[pos+t.size()/2] = 'X';
+             ans.pb(pos+t.size()/2);
+          }
+      }
+      cout<<ans.size()<<endl;
+      for(auto x:ans)
+        cout<<x+1<<endl;;
+pos id the first index if we found out any of the 3 "strings"
+we change the mid value to 'x' for eg twone becomes twXne;
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1415/C
+int n,p,k;
+      cin>>n>>p>>k;
+      p--;
+      vector<int>v(n);
+      string s;
+      cin>>s;
+      int x,y;
+      cin>>x>>y;
+      int ans = INT_MAX;
+      for(int i=n-1; i>=0; i--)
+      {
+          v[i] = !(s[i]=='1');
+          if(i+k<n)
+            v[i]+=v[i+k];
+          if(i>=p)
+          {
+             ans = min(ans, (i-p)*y+v[i]*x);
+          }
+      }
+      cout<<ans<<endl;
+DIFFERENT SYNTAX "v[i] = !(s[i]=='1')";
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1519/problem/D
+vector<int> a(n), b(n);
+      for(int i=0; i<n; i++)
+        cin>>a[i];
+      for(int i=0; i<n; i++)
+        cin>>b[i];
+
+      int ans = 0;
+      int maxi;
+      for(int i=0; i<n; i++)
+        ans+=a[i]*b[i];
+
+      maxi=ans;
+      for(int i=0; i<n; i++)
+      {
+          int total=ans;
+          int l=i-1;
+          int r=i+1;
+          while(l>=0 && r<n)
+          {
+             total-=(a[l]*b[l] + a[r]*b[r]);
+             total+=(a[l]*b[r] + a[r]*b[l]);
+             maxi=max(maxi,total);
+             l--;
+             r++;
+          }
+      }
+      for(int i=0; i<n; i++)
+      {
+         int total=ans;
+         int l=i;
+         int r=i+1;
+         while(l>=0 && r<n)
+         {
+            total-=(a[l]*b[l] + a[r]*b[r]);
+            total+=(a[l]*b[r] + a[r]*b[l]);
+            maxi=max(maxi,total);
+            l--;
+            r++;
+         }
+      }
+      cout<<maxi;
+#2 POINTERS
+
+
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/contest/1389/problem/C
+string s;
+      cin>>s;
+      int n = s.size();
+      int f[10] = {};
+      for(int i=0; i<n; i++)
+        f[s[i]-'0']++;
+
+      int max_frequency = 0;
+      for(int i=0; i<10; i++)
+        max_frequency = max(max_frequency, f[i]);
+
+      int ans = s.size()-max_frequency;
+      for(int i=0; i<10; i++)
+      {
+          for(int j=0; j<10; j++)
+          {
+              bool flag = true;
+              int t=0;
+              for(int k=0; k<n; k++)
+              {
+                 if(flag && s[k]=='0'+i)
+                 {
+                    flag=false;
+                    t++;
+                 }
+                 else if(!flag && s[k]=='0'+j)
+                 {
+                    flag=true;
+                    t++;
+                 }
+              }
+              if(t%2==1)
+                t--;
+
+              ans=min(ans,n-t);
+          }
+      }
+      cout<<ans<<endl;
+      s[k] == '0' + i
+Is used for comparison
+'0' + i is the integer i represented as char
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/455/A
+int n;
+      cin>>n;
+      int maxi = -1;
+      vector<int> a(100001, 0);
+      for(int i=0; i<n; i++)
+      {
+          int x;
+          cin>>x;
+          a[x]++;
+          maxi = max(maxi, x);
+      }
+      vector<int> b(100001, 0);
+      b[0]=0;
+      b[1]=a[1];
+      for(int i=2; i<100001; i++)
+      {
+          b[i] = max(b[i-1], b[i-2]+(a[i]*i));
+      }
+      cout<<b[maxi];
+#DP
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/580/C
+int n,k;
+vector<int> c, vis;
+vector<vector<int>> g;
+int a=0;
+
+
+void dfs(int starting_vertex, int m)
+{
+    vis[starting_vertex]=1;
+
+    if(m>k)
+      return;
+
+    bool f=true;
+    for(auto x:g[starting_vertex])
+    {
+       if(!vis[x])
+       {
+          f=false;
+          dfs(x, m*c[x] + c[x]);
+       }
+    }
+    if(f)
+      a++;
+
+}
+
+
+
+void solve() {
+
+    cin>>n>>k;
+    c.resize(n);
+    for(int i=0; i<n; i++)
+      cin>>c[i];
+
+    g.resize(n);
+    for(int i=0; i<n-1; i++)
+    {
+       int a,b;
+       cin>>a>>b;
+       a--, b--;
+
+       g[a].pb(b);
+       g[b].pb(a);
+    }
+    vis.resize(n, 0);
+
+    dfs(0, c[0]);
+    cout<<a;
+}
+
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/550/C
+string s;
+    cin>>s;
+
+    for(char c:s)
+    {
+       if((c-'0')%8==0)
+       {
+          yes;
+          cout<<c<<endl;
+          return;
+       }
+    }
+    for(int i=0; i<s.size(); i++)
+    {
+       for(int j=i+1; j<s.size(); j++)
+       {
+          int f = (s[i]-'0')*10 + (s[j]-'0');
+          if(f%8==0)
+          {
+              yes;
+              cout<<f<<endl;
+              return;
+          }
+        }
+    }
+    for(int i=0; i<s.size(); i++)
+    {
+       for(int j=i+1; j<s.size(); j++)
+       {
+          for(int k=j+1; k<s.size(); k++)
+          {
+              int f = (s[i]-'0')*100 + (s[j]-'0')*10 + (s[k]-'0');
+              if(f%8==0)
+              {
+                 yes;
+                 cout<<f<<endl;
+                 return;
+              }
+          }
+       }
+    }
+    no;
+taking out numbers from a string and checking if they are divisible by 8;
+#MATH
+#NUMBERTHEORY
+
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1399/D
+int n;
+    cin>>n;
+    string s;
+    cin>>s;
+    stack<int> one, zer;
+    vector<int> a(n);
+    int numpile = 1;
+    for(int i=0; i<n; i++)
+    {
+        if(s[i]=='0')
+        {
+           if(one.empty())
+           {
+              zer.push(numpile);
+              a[i]=numpile;
+              numpile++;
+           }
+           else
+           {
+              int k=one.top();
+              one.pop();
+              a[i]=k;
+              zer.push(k);
+           }
+        }
+        else
+        {
+            if(zer.empty())
+            {
+               one.push(numpile);
+               a[i]=numpile;
+               numpile++;
+            }
+            else
+            {
+                int k = zer.top();
+                zer.pop();
+                a[i]=k;
+                one.push(k);
+            }
+        }
+    }
+    cout<<numpile-1<<endl;
+    for(auto x:a)
+      cout<<x<<" ";
+
+    cout<<endl;
+
+
+
+
+
+
+
+
+
+https://codeforces.com/problemset/problem/1538/F
+int solve(int x)
+{
+    int r=0;
+    while(x>0)
+    {
+        r+=x;
+        x/=10;
+    }
+    return r;
+}
+
+
+
+void solve() {
+
+    int n,m;
+    cin>>n>>m;
+
+    int a=solve(n);
+    int b=solve(m);
+
+    cout<<abs(a-b)<<endl;
+  }
+#MATH
